88. Link between pages and recognize that a user can see videos(Section 7: Video subscription service using PayPal Subscription & Symfony PHP framework)
Let's make links between pages because for now it doesn't work.

And let's also fix this during this lecture because I have two informations here.

The video for the subscriber and the message for the user that is not a subscriber or he's not logged

into the application.

These two informations cannot exist together.

So we will fix this in this lecture.

First of all let's open subscription controller and create private function can watch videos

and if this get user which means that if the user is logged into the application then return true or

otherwise return false of course for now there is a missing logic when it comes to subscription functionality.

We will do this in the future.

For now let's live it like this.

And in the videos method let's use this can watch videos.

Private function if it returns true then it can watch videos equals true or otherwise can watch videos

equals false and let's assign countless videos of variable to the view.

Like so using built in a compact function next Open videos HDMI wi fi and if candlelit videos then a

user is able to see video or the device will show him this message.

And let's also change this old link using the power function and the argument pricing itself will change

here.

Bath videos

cease all of the changes and to refresh the page in order to watch videos you must be logged in and

have subscription valid links.

Also work OK.

Of course I not have a logging system yet.

So that's why logging doesn't work.

I haven't even any are out for this or will do this in the next lecture.