90. Update User entity(Section 7: Video subscription service using PayPal Subscription & Symfony PHP framework)
Before we make a registration system we need to customize the user entity a little bit.

I mean we need to add some properties to it open to council and run command and council make entity

it is a little bit confusing but the same command is used for creating as well as updating existing

entity and type the name user

and we see your entity already exists.

And your property name.

Name because our user has to have some name and this is going to be string type filled.

So let's enter a default.

String as well as default 255 land can this field be known in the database.

Yes that's another property subscription status.

And there's going to be boolean type field and is still not in the database yes because if a user registers

at the beginning he is not a subscriber because subscription status column in the database will be used

to check whether a user is a subscriber or not at another property agreement I.D. agreement idea can

be used to show the details of the subscription for example whether it is active or not expired or not.

And so on.

This is going to string tied into 255 land

can this still be not in the database yes

and add another property just hit enter delete this command prompt but let's make this comment again

because I forgot about the additional property or Ada in order to identify the subscriber.

This is going to be strength type and two hundred fifty five land.

Can this still be not in the database.

Yes.

And hit enter to live.

This command prompt okay a changed entity and because of that we need to run the following command and

council doctrine migrations if it is going to change our migration file according to what we updated

in user entity and after that apply all the changes in the database by running the following command

mean council doctrine migrations migrate IBS here and enter and as you can see my database and tables

were changed

as you can see user and he was updated new properties were added as well as methods and so on.

Let's also add here the unique entity object like this unique entity fields email message there is already

an account of this email in order to make email call them in database unique but if you change your

entity again unique to AROUND THE being Council directed migrations this one more time as well as been

consulted structuring migrations great to apply changes in the database.