91. Make register functionality in Symfony application(Section 7: Video subscription service using PayPal Subscription & Symfony PHP framework) In the previous lecture we updated user entity here for additional properties and methods for example
get subscription status which is going to return the value of the subscription status column in user

table as well as set subscription status to set a new value for that column subscription status value

will be used here in private function can watch videos in the subscription controller.

So now we can add another condition if user is logged in and subscription status column is equal to

true then this user is allowed to watch videos.

So now this is the full condition for our subscription application

and we use can watch of videos method in the videos.

Method

and this is the condition in the view file based on which we show videos or in that.

Okay.

We have a logging system but there is no registration system so let's do this now open the console and

to run the following command have been console make registration form

and the question do you want to automatically authenticate the user after registration it default.

Yes here enter three files were created.

Registration Form type registration controller as well as register of view.

This is all we need to do to have a fully working registration system but I don't have register a link

yet.

So let's say that we will have register a link next to a sign in button so let's open a security folder

and log in each to a melting view and below the link to register this real name of course exists because

if you open a registration controller which was automatically generated we have such a crowd here and

the method that handles this round okay.

Let's save this and refresh the page and we have the link to a register.

Let's click it.

We have registration of forum but it looks ugly but in symphony it is very easy to customize.

Forums this open conflict directory packages and weekly annual file.

Let's add here a one configuration entry form themes bootstrap for layout and now if we refresh the

page we have bootstrap for styled HDMI form.

For some reason I don't see a register button.

So let's fix that.

Open the register view and add here.

See asset class for example between success now with loads better Okay let's try to register a user

you see that the front end validation works.

Let's see if the back end validation works correctly.

Let's try to register with some incorrect email value.

But it looks like we called somewhere in the view logo trout which doesn't exist yet.

So let's create a logo with the method and the root insecurity controller but all we need to do here

is the real definition the old name and the method the method can be empty Symphony will handle logo

process for me.

We can throw here exception.

We also need to configure logout insecurity ammo file let's open that file and add logo section.

Remember about two tabulations and pass a BP logo out and target videos.

Which means that after logging out we will be redirected to the videos road.

Okay let's save this and try one more time with incorrect email address.

And I was registered to with incorrect email address so let's fix that logo to work as we can see.

So as you can see after creating registration functionality we don't have validation of the email fields

out of the box.

So let's fix that open the registration form time and for the password.

We have validation as you can see.

Let's create validation for the email field email type class import that class namespace

and constrains your email

important the namespace for this and the message.

Please enter a valid email say this and try to one more time or register with invalid email address.

And this time we see the message Please enter a valid email.

Our validation works the food to correct email address and register and I was not only registered but

also automatically logged in to the application after registration process.

As you can see we are not able to see videos according to the logic here.

I am locked in bad I am not a subscriber yet because the default value of the subscription status column

is is null as defined here.