93. Activate subscription plan(Section 7: Video subscription service using PayPal Subscription & Symfony PHP framework)
In the previous lecture we created a subscription plan but we need to activate this plan if we want

our subscribers to subscribe to it.

Get rid of this and instead this Activate plan created plan.

Private function Activate plan and try catch block new people API patch.

What I am doing now is of course based on the code available on this website.

If we go to push B sample code building agreements we have great plan and update activate Plan code

and based on this.

Now I am going to activate creative plan users may subscribe only to active plans.

New people come on people model and this is going to be the new value for the state of the plan.

Active in Jason format.

Set operational replace set bar slash which is a little bit confusing.

This is not any URL or something like that.

This is the place in Jason File where we want to update some value.

In this case this is the route.

This is Sir Jason patch notation standard and set value parts for request.

Add a bunch patch and created plan update patch request in the CPA context in order to not repeat the

IPA context data here.

Let's also add these which is the class variables since now and I will assign to this class of variables

something in the moment

and return created.

In case some error is less printed them OK.

Now let's fix this.

Let's get rid of this from here

and create new folder utilize in the SLC directory and inside of it.

New file PayPal Api credential striped namespace ATP utilize

and tried.

People API credentials right and here I'm going to store API context for rest api authentication public

function set credentials.

In this API context and here I write exactly what was written in the controller people client idea and

people client secret from Eon v file let's import here the namespace for that right.

And to use it here.

Use statement inside the class works like include for example.

In other words all the methods from the tried are available here in this class and says create constructor

and call set credentials right from the right since now these API contexts variable is available for

this controller.

Now if a great alarm it is going to be not only created but also activated and ready to subscribe to.