94. List all subscription plans(Section 7: Video subscription service using PayPal Subscription & Symfony PHP framework)
After the administrator created subscription plan and activated it we can list all the created and activated

plans on the website.

And now I am going to do this based on this code open PayPal admin controller and public functions show

plans.

Some parameters what the show and how you can show all or only created active inactive and so on.

I am going to show all plans which are created and activated together and PayPal Api.

All forums and this API context and brain to our plan A Story Of course this method is going to be initialized

by the Round so let's define it.

For example show plans as the path and name path is going to be preceded of course by the admin which

is defined here.

Okay let's save this and go to the website

and their admin show plus your URL and press enter and we have listed all the plans

active and active and this one is just only created because this one was created when there was no activation

code yet the plan like this.

A user is not allowed to subscribe only to plans that are active.